ENH: A point-in-time restricted list with restrictions store in memory
An ABC RLManager defines a group of restricted list managers responsible
for returning restriction information for sids on certain dts. An
InMemoryRLManager is a point-in-time manager, with all restrictions
and their dates passed in upon instantiation. A StaticRestrictedList
takes a list of sids, restricting them at all dates. A
RestrictionsController returns the restriction information for a group
of restriction list managers on aggregate